refactor: change Archive to Article
