+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BrowserRouter, Switch } from 'react-router-dom'
 import DefaultLayout from './layouts/DefaultLayout'
 import Home from './pages/Home'
-import Archive from './pages/Archive'
+import Archive from './pages/Articles'
 import ArchiveArticle from './pages/ArchiveArticle'
 import Error from './pages/Error'
 import { routePath } from './constants'
@@ -13,14 +13,11 @@ function App() {
       <Switch>
         <DefaultLayout path={routePath.HOME} component={Home} exact={true} />
         <DefaultLayout
-          path={routePath.ARCHIVES}
+          path={routePath.ARTICLES}
           component={Archive}
           exact={true}
         />
-        <DefaultLayout
-          path={routePath.ARCHIVE_ARTICLE}
-          component={ArchiveArticle}
-        />
+        <DefaultLayout path={routePath.ARTICLE} component={ArchiveArticle} />
         <DefaultLayout path="*" component={Error} />
       </Switch>
     </BrowserRouter>
