+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export interface Article {
+  id: string
   title: string
   imgUrl: string
   href: string
